Internal Functions

SchumakerSpline.Schumaker

SchumakerSpline.Schumaker_ExtrapolationSchemes

SchumakerSpline.evaluate

SchumakerSpline.evaluate_integral

SchumakerSpline.find_derivative_spline

SchumakerSpline.find_optima

SchumakerSpline.find_roots

SchumakerSpline.get_crossover_in_interval

Main Struct

SchumakerSpline.Schumaker — Type

Schumaker(x::Array{T,1},y::Array{T,1} ; gradients::Union{Missing,Array{T,1}} = missing, extrapolation::Schumaker_ExtrapolationSchemes = Curve,
              left_gradient::Union{Missing,T} = missing, right_gradient::Union{Missing,T} = missing)
Schumaker(x::Array{Int,1},y::Array{T,1} ; gradients::Union{Missing,Array{T,1}} = missing, extrapolation::Schumaker_ExtrapolationSchemes = Curve,
              left_gradient::Union{Missing,T} = missing, right_gradient::Union{Missing,T} = missing)
Schumaker(x::Array{Date,1},y::Array{T,1} ; gradients::Union{Missing,Array{T,1}} = missing, extrapolation::Schumaker_ExtrapolationSchemes = Curve,
              left_gradient::Union{Missing,T} = missing, right_gradient::Union{Missing,T} = missing)

Creates a Schumaker spline.

Takes

x - A vector of x coordinates.

y - A vector of y coordinates.

extrapolation (optional) - This should be Curve, Linear or Constant specifying how to interpolate outside of the sample domain. By default it is Curve which extends out the first and last quadratic curves. The other options are Linear which extends the line (from first and last curve) out from the first and last point and Constant which extends out the y value at the first and last point.

gradients (optional)- A vector of gradients at each point. If not supplied these are imputed from x and y.

left_gradient - The gradient at the lowest value of x in the domain. This will override the gradient imputed or submitted in the gradients optional argument (if it is submitted there)

right_gradient - The gradient at the highest value of x in the domain. This will override the gradient imputed or submitted in the gradients optional argument (if it is submitted there)

Returns

A Schumaker object which details the spline. This object can then be evaluated with evaluate or evaluate_integral.

SchumakerSpline.evaluate — Function

Evaluates the spline at a point. The point can be specified as a Real number (Int, Float, etc) or a Date. Derivatives can also be taken.

Takes

spline - A Schumaker type spline

PointToExamine - The point at which to evaluate the integral

derivative - The derivative being sought. This should be 0 to just evaluate the spline, 1 for the first derivative or 2 for a second derivative.

Higher derivatives are all zero (because it is a quadratic spline). Negative values do not give integrals. Use evaluate_integral instead.

Returns

A value of the spline or appropriate derivative in the same format as specified in the spline.

SchumakerSpline.evaluate_integral — Function

Estimates the integral of the spline between lhs and rhs. These end points can be input as Reals or Dates.

Takes

spline - A Schumaker type spline

lhs - The left hand limit of the integral

rhs - The right hand limit of the integral

Returns

A Float64 value of the integral.

Extrapolation Schemes

SchumakerSpline.Schumaker_ExtrapolationSchemes — Type

This creates an enum which details how extrapolation from the interpolation domain should be done.

Missing docstring.

Missing docstring for Curve. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Linear. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Constant. Check Documenter's build log for details.

Working with Splines

SchumakerSpline.find_derivative_spline — Function

findderivativespline(spline::Schumaker) Returns a SchumakerSpline that is the derivative of the input spline

SchumakerSpline.find_roots — Function

findroot(spline::Schumaker; rootvalue::T = 0.0) Finds roots - This is handy because in many applications schumaker splines are monotonic and globally concave/convex and so it is easy to find roots. Here root_value can be set to get all points at which the function is equal to the root value. For instance if you want to find all points at which the spline has a value of 1.0.

SchumakerSpline.find_optima — Function

find_optima(spline::Schumaker) Finds optima - This is handy because in many applications schumaker splines are monotonic and globally concave/convex and so it is easy to find optima.

SchumakerSpline.get_crossover_in_interval — Function

get_crossover_in_interval(s1::Schumaker{T}, s2::Schumaker{R}, interval::Tuple{U,U}) where T<:Real where R<:Real where U<:Real

Finds the point at which two schumaker splines cross over each other within a single interval. This is not exported.

Missing docstring.

Missing docstring for get_intersection_points. Check Documenter's build log for details.

Missing docstring.

Missing docstring for reshape_values. Check Documenter's build log for details.

Missing docstring.

Missing docstring for splice_splines. Check Documenter's build log for details.

Two dimensional splines

Missing docstring.

Missing docstring for Schumaker2d. Check Documenter's build log for details.

Plotting of splines

Missing docstring.

Missing docstring for plot. Check Documenter's build log for details.